refactor: Review rating double → int 변경, ReviewDto author 객체로 변경, ReviewMapper toDto 직접 변환으로 교체

+++ src/main/java/com/codeit/mpl/domain/review/dto/request/ReviewCreateRequest.java
@@ -11,5 +11,5 @@ public record ReviewCreateRequest(
     @NotBlank String text,      // ← @NotNull에서 @NotBlank로 변경
     @NotNull
     @DecimalMin("0.0")
-    @DecimalMax("5.0") Double rating
+    @DecimalMax("5.0") int rating
 ) {}
